src/server/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
grpc-interceptor ~= 0.12.0
python-dotenv==0.21.1
openai
